requirements.txt
# Core Dependencies
# These are required for the project to run
PyYAML

# AI API Dependencies
# 用户可以选择使用OpenAI或Google Gemini API
openai>=1.0.0
google-genai

# Development Dependencies
# These are used for testing and development
pytest
pytest-mock

# Refactoring Dependencies
pydantic
json-repair
python-dotenv
requests
colorama
fastapi
uvicorn[standard]
python-multipart

# Phonetic Matching Dependencies
pypinyin
pykakasi
jamo
python-Levenshtein
